# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/units/{id}": {
      "get": {
        "operationId": "UnitController_getUnitById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Unit UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get unit by ID",
        "tags": [
          "Unit Management"
        ]
      },
      "put": {
        "operationId": "UnitController_updateUnit",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Unit UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUnitDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a unit",
        "tags": [
          "Unit Management"
        ]
      },
      "delete": {
        "operationId": "UnitController_deleteUnit",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Unit UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a unit",
        "tags": [
          "Unit Management"
        ]
      }
    },
...
    "/master-unit-types": {
      "get": {
        "operationId": "MasterUnitTypeController_findAll",
        "parameters": [
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "Sort order (asc/desc)",
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Field to sort by",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search keyword",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page (default: 10)",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (default: 1)",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "isSuccess": {
                          "type": "boolean",
                          "example": true
                        },
                        "message": {
                          "type": "string",
                          "example": "Success"
                        },
                        "data": {
                          "properties": {
                            "items": {
                              "type": "array",
                              "items": {
                                "$ref": "#/components/schemas/MasterUnitTypeDto"
                              }
                            },
                            "pagination": {
                              "properties": {
                                "page": {
                                  "type": "number",
                                  "example": 1
                                },
                                "limit": {
                                  "type": "number",
                                  "example": 10
                                },
                                "totalItems": {
                                  "type": "number",
                                  "example": 100
                                },
                                "totalPages": {
                                  "type": "number",
                                  "example": 10
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all master unit types",
        "tags": [
          "Master Unit Type"
        ]
      },
      "post": {
        "operationId": "MasterUnitTypeController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMasterUnitTypeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MasterUnitTypeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create new master unit type",
        "tags": [
          "Master Unit Type"
        ]
      }
    },
    "/master-unit-types/{id}": {
      "get": {
        "operationId": "MasterUnitTypeController_findById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MasterUnitTypeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get master unit type by ID",
        "tags": [
          "Master Unit Type"
        ]
      },
      "put": {
        "operationId": "MasterUnitTypeController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMasterUnitTypeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MasterUnitTypeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update master unit type",
        "tags": [
          "Master Unit Type"
        ]
      },
      "delete": {
        "operationId": "MasterUnitTypeController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Master unit type successfully deleted"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete master unit type",
        "tags": [
          "Master Unit Type"
        ]
      }
    }
  },
  "info": {
    "title": "SIM PROKER API",
    "description": "Program Kerja Information System — REST API Documentation.\n\nThis service manages program planning, activity management, output tracking, progress monitoring, evidence management, and approval workflows.\n\n**Authentication**: All endpoints require a valid JWT Bearer token from Auth Service.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter the JWT token obtained from Auth Service"
      }
    },
    "schemas": {
...
      "UpdateUnitDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Unit A"
          },
          "description": {
            "type": "string",
            "example": "Description for Unit A"
          }
        }
      },
...
      "MasterUnitTypeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440000",
            "description": "Master Unit Type UUID"
          },
          "name": {
            "type": "string",
            "example": "Dokumen",
            "description": "Name of the unit type"
          },
          "type": {
            "type": "string",
            "enum": [
              "NUMBER",
              "FILE",
              "TEXT"
            ],
            "example": "FILE",
            "description": "Type of the unit (NUMBER, FILE, TEXT)"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-01T00:00:00.000Z",
            "description": "Creation timestamp"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-01T00:00:00.000Z",
            "description": "Update timestamp"
          }
        },
        "required": [
          "id",
          "name",
          "type",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateMasterUnitTypeDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Dokumen",
            "description": "Name of the unit type"
          },
          "type": {
            "type": "string",
            "enum": [
              "NUMBER",
              "FILE",
              "TEXT"
            ],
            "example": "FILE",
            "description": "Type of the unit (NUMBER, FILE, TEXT)"
          }
        },
        "required": [
          "name",
          "type"
        ]
      },
      "UpdateMasterUnitTypeDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Dokumen",
            "description": "Name of the unit type"
          },
          "type": {
            "type": "string",
            "enum": [
              "NUMBER",
              "FILE",
              "TEXT"
            ],
            "example": "FILE",
            "description": "Type of the unit (NUMBER, FILE, TEXT)"
          }
        }
      }
    }
  }
}